enterprise-grade-rag/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
rank_bm25==0.2.2
scikit-learn==1.5.1
numpy==1.26.4
scipy==1.11.4
transformers==4.43.3
sentence-transformers==3.0.1
torch>=2.1.0
requests==2.32.3
pydantic==2.8.2
faiss-cpu==1.8.0
cachetools==5.3.3
prometheus-client==0.20.0
ragas==0.1.14
langsmith==0.1.99
datasets==2.20.0
